fix(policy-titles-admin): move ApproveBlockedError out of 'use server' actions (Next forbids non-async exports)
